OpenRouter
Moved away from the idea of supporting every provider individually because fuck me dead, man… we can still do that when we have a lot of time because it would be cheaper, but there is no way in hell I will use all major providers before this to come close to the model support.

We now have selected OpenAI, Anthropic, Google, Mistral and DeepSeek models. This list will not be significantly expanded, instead, these will be curated models with the user being able to browse the full list in a more convoluted UI (by design because… long lists are convoluted).

Will also soon group the providers properly in the model picker because goddamn.