Removing deleted agents/chats from other devices
If an agent or device does not exist in the cloud, it will be deleted. Will see how we handle merge conflicts in the future, then… it shouldn't an issue because if a device is capable of chatting, it is online anyways and would be able to push new agents so data loss should not happen.